Move expr visitor to common_sql package
