+++ common/functions/Cargo.toml
@@ -16,9 +16,11 @@ common-arrow = { path = "../arrow" }
 common-datavalues = { path = "../datavalues" }
 common-exception = { path = "../exception" }
 common-io = { path = "../io" }
+common-sql = { path = "../sql" }
 
 # Github dependencies
 bincode = { git = "https://github.com/datafuse-extras/bincode", rev = "fd3f9ff" }
+sqlparser = { git = "https://github.com/datafuse-extras/sqlparser-rs", rev = "b42ff01" }
 
 # Crates.io dependencies
 dyn-clone = "1.0.4"
@@ -44,6 +46,7 @@ bstr = "0.2.17"
 uuid = { version = "0.8", features = ["v4"] }
 strength_reduce = "0.2.3"
 num-traits = "0.2"
+regex = "1.5.4"
 
 [dev-dependencies]
 bumpalo = "3.8.0"
