+++ common/sql/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "common-sql"
+version = "0.1.0"
+authors = ["Databend Authors <opensource@datafuselabs.com>"]
+license = "Apache-2.0"
+publish = false
+edition = "2021"
+
+[lib]
+doctest = false
+test = false
+
+[dependencies] # In alphabetical order
+# Workspace dependencies
+common-exception = { path = "../exception" }
+
+# Github dependencies
+sqlparser = { git = "https://github.com/datafuse-extras/sqlparser-rs", rev = "b42ff01" }
+
+# Crates.io dependencies
+
+[dev-dependencies]
