+++ common/sql/src/lib.rs
@@ -12,21 +12,4 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod common;
-mod exec;
-mod metrics;
-pub mod optimizer;
-pub mod parser;
-mod plan_parser;
-mod planner;
-mod sql_common;
-mod sql_parser;
-mod sql_statement;
-pub mod statements;
-
-pub use common::*;
-pub use plan_parser::PlanParser;
-pub use planner::*;
-pub use sql_common::SQLCommon;
-pub use sql_parser::DfParser;
-pub use sql_statement::*;
+pub mod expr;
