+++ query/Cargo.toml
@@ -49,6 +49,7 @@ common-metrics = { path = "../common/metrics" }
 common-planners = { path = "../common/planners" }
 common-streams = { path = "../common/streams" }
 common-tracing = { path = "../common/tracing" }
+common-sql = { path = "../common/sql" }
 
 # Github dependencies
 clickhouse-rs = { git = "https://github.com/datafuse-extras/clickhouse-rs", rev = "fe0406b" }
@@ -102,7 +103,6 @@ tonic = "0.6.2"
 uuid = { version = "0.8.2", features = ["serde", "v4"] }
 walkdir = "2.3.2"
 parquet-format-async-temp = "0.2.0"
-regex = "1.5.4"
 
 [dev-dependencies]
 criterion = "0.3.5"
