+++ query/src/sql/statements/statement_create_udf.rs
@@ -15,12 +15,12 @@
 use std::sync::Arc;
 
 use common_exception::Result;
+use common_functions::udfs::UDFParser;
 use common_meta_types::UserDefinedFunction;
 use common_planners::CreateUDFPlan;
 use common_planners::PlanNode;
 use common_tracing::tracing;
 
-use crate::functions::UDF;
 use crate::sessions::QueryContext;
 use crate::sql::statements::AnalyzableStatement;
 use crate::sql::statements::AnalyzedResult;
@@ -37,8 +37,8 @@ pub struct DfCreateUDF {
 impl AnalyzableStatement for DfCreateUDF {
     #[tracing::instrument(level = "info", skip(self, _ctx), fields(ctx.id = _ctx.get_id().as_str()))]
     async fn analyze(&self, _ctx: Arc<QueryContext>) -> Result<AnalyzedResult> {
-        let mut udf = UDF::default();
-        udf.parse_definition(self.definition.as_str())?;
+        let mut udf_parser = UDFParser::default();
+        udf_parser.parse_definition(self.definition.as_str())?;
 
         Ok(AnalyzedResult::SimpleQuery(Box::new(PlanNode::CreateUDF(
             CreateUDFPlan {
